# static/swagger/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0.0"
  title: Mechanic Workshop API
  description: API documentation for Customers, Mechanics, Service Tickets, and Inventory.
# API server configuration specifying host, base path, and HTTP scheme
host: localhost:5000
basePath: /api
schemes:
  - http
# Security definition for JWT Bearer token authentication via Authorization header
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: >
      JWT Token authentication. Use 'Bearer <your-token>' format.
# Tags categorizing API endpoints by resource type
...
paths:
...
  /mechanics:
    post:
      tags:
        - Mechanics
      summary: Create Mechanic
      description: "Registers a new mechanic account."
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/MechanicInput'
            example:
              name: "Emily Smith"
              email: "emily@example.com"
              password: "securepassword123"
              specialty: "Engine Repair"
              phone: "+1-555-234-5678"
              address: "12 Oak Street, Metropolis, NY"
              salary: 45000
      responses:
        201:
          description: "Mechanic created successfully"
          schema:
            $ref: '#/definitions/Mechanic'
          examples:
            application/json:
              id: 1
              name: "Emily Smith"
              email: "emily@example.com"
              specialty: "Engine Repair"
              phone: "+1-555-234-5678"
              address: "12 Oak Street, Metropolis, NY"
              salary: 45000
# Retrieve a paginated list of all registered mechanics (requires authentication)
  /mechanics/:
    get:
      tags:
        - Mechanics
      summary: Find All Mechanics (Paginated)
      description: Returns a paginated list of mechanics.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          type: integer
          required: false
          description: "Page number (default: 1)"
        - name: per_page
          in: query
          type: integer
          required: false
          description: "Results per page (default: 10)"
      responses:
        200:
          description: A paginated list of mechanics.
          schema:
            type: object
            properties:
              mechanics:
                type: array
                items:
                  $ref: '#/definitions/Mechanic'
              total:
                type: integer
              pages:
                type: integer
              current_page:
                type: integer
# Authenticate a mechanic and return a JWT token
  /mechanics/login:
    post:
      tags:
        - Mechanics
      summary: Mechanic Login (Token Auth)
      description: "Logs in a mechanic and returns a JWT token along with mechanic details."
      parameters:
        - in: body
          name: credentials
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
            example:
              email: "emily@example.com"
              password: "supersecure123"
      responses:
        200:
          description: "Login successful"
          schema:
            type: object
            properties:
              mechanic_id:
                type: integer
              mechanic_name:
                type: string
              message:
                type: string
              token:
                type: string
          examples:
            application/json:
              mechanic_id: 35
              mechanic_name: "Samantha Lee"
              message: "Login successful"
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkI..."
# Retrieve mechanic details by their ID (requires authentication)
  /mechanics/{id}:
    get:
      tags: [Mechanics]
      summary: Find Mechanic by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          schema:
            $ref: '#/definitions/Mechanic'
# Update a mechanic's information by their ID (requires authentication)
    put:
      tags:
        - Mechanics
      summary: Update Mechanic by ID
      description: "Updates an existing mechanic account."
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: ID of the mechanic to update
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/MechanicInput'
            example:
              name: "Emily Smith"
              email: "emily@example.com"
              password: "newpassword123"
              specialty: "Engine Repair"
              phone: "+1-555-234-5678"
              address: "12 Oak Street, Metropolis, NY"
              salary: 45000
      responses:
        200:
          description: "Mechanic updated successfully"
          schema:
            $ref: '#/definitions/Mechanic'
          examples:
            application/json:
              id: 1
              name: "Emily Smith"
              email: "emily@example.com"
              specialty: "Engine Repair"
              phone: "+1-555-234-5678"
              address: "12 Oak Street, Metropolis, NY"
              salary: 45000
# Delete a mechanic account by their ID (requires authentication)
    delete:
      tags: [Mechanics]
      summary: Delete Mechanic by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        204:
          description: Mechanic deleted successfully
# Retrieve a list of top-performing mechanics (requires authentication)
  /mechanics/top:
    get:
      tags: [Mechanics]
      summary: Find Top Mechanics
      security:
        - BearerAuth: []
      responses:
        200:
          description: Top mechanics
          schema:
            type: array
            items:
              $ref: '#/definitions/Mechanic'
# Create a new service ticket (requires authentication)
...
definitions:
...
  Mechanic:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      specialty:
        type: string
      salary:
        type: number
    required:
      - name
      - email
      - phone
      - specialty
      - salary
# Input model for creating or updating a mechanic with required fields
  MechanicInput:
    type: object
    required: [name, email, password, address, phone, specialty, salary]
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
      address:
        type: string
      phone:
        type: string
      specialty:
        type: string
      salary:
        type: number
# ServiceTicket object model representing a service ticket's details
...